+++ example/lib/main.dart
@@ -251,13 +251,13 @@ const inUseEmails = ['johndoe@email.com', 'john@email.com'];
 Future<Map<String, dynamic>> _uniqueEmail(AbstractControl control) async {
   final error = {'unique': false};
 
-  final emailAlreadyUsed = await Future.delayed(
+  final emailAlreadyInUse = await Future.delayed(
     Duration(seconds: 5), // a delay to simulate a time consuming operation
     () => inUseEmails.contains(control.value),
   );
 
-  if (emailAlreadyUsed) {
-    control.touch();
+  if (emailAlreadyInUse) {
+    control.markAsTouched();
     return error;
   }